labs/agents_cli/requirements.txt
json
os
Groq
load_dotenv
pydantic
ast